Change LoadCommand to ImportCommand
