+++ src/main/java/seedu/address/logic/parser/ImportCommandParser.java
@@ -2,23 +2,23 @@ package seedu.address.logic.parser;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.LoadCommand.MESSAGE_USAGE;
+import static seedu.address.logic.commands.ImportCommand.MESSAGE_USAGE;
 
-import seedu.address.logic.commands.LoadCommand;
+import seedu.address.logic.commands.ImportCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 //@@author k-l-a
 /**
- * Parses input arguments and create a new LoadCommand Object.
+ * Parses input arguments and create a new ImportCommand Object.
  */
-public class LoadCommandParser implements Parser<LoadCommand> {
+public class ImportCommandParser implements Parser<ImportCommand> {
 
     /**
-     * Parses the given (@code String) in the context of a LoadCommand.
+     * Parses the given (@code String) in the context of a ImportCommand.
      * @param args
-     * @return LoadCommand Object for execution
+     * @return ImportCommand Object for execution
      */
-    public LoadCommand parse(String args) throws ParseException {
+    public ImportCommand parse(String args) throws ParseException {
         requireNonNull(args);
 
         String trimmedArgs = args.trim();
@@ -26,6 +26,6 @@ public class LoadCommandParser implements Parser<LoadCommand> {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, MESSAGE_USAGE));
         }
 
-        return new LoadCommand(trimmedArgs);
+        return new ImportCommand(trimmedArgs);
     }
 }
